制限+4km/hでハーフブレーキ(減速定数:18)
コピーコンストラクタ追加
spec手渡し
渡す制限速度をfloatに変更

+++ R-ATC/R_ATC.h
@@ -27,7 +27,7 @@ private:
 	//     vehicle status
 	// return : uint16_t
 	//     output speed limit
-	uint16_t calclateSpeed(VehicleState state);
+	float calclateSpeed(VehicleState state);
 
 	// calclate brake notch
 	// arg : VehicleState state
@@ -38,7 +38,7 @@ private:
 	//     ATC parameter
 	// return : uint16_t
 	//     output brake notch
-	uint16_t calclateBrake(VehicleState state, uint16_t speed, uint8_t param = 0);
+	uint16_t calclateBrake(VehicleState state, float speed, uint8_t param = 0);
 
 public:
 	// constructor
